backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
psycopg2-binary>=2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
stripe==10.5.0
pydantic[email]==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
fpdf2==2.8.2
